+++ reward_kit/mcp/execution/__init__.py
@@ -1,0 +1,14 @@
+"""
+MCP Execution Framework
+
+This module handles policy execution, tool calling, and rollout coordination.
+"""
+
+from .policy import FireworksPolicy, LLMBasePolicy
+from .rollout import RolloutManager
+
+__all__ = [
+    "LLMBasePolicy",
+    "FireworksPolicy",
+    "RolloutManager",
+]
