+++ reward_kit/mcp/session/__init__.py
@@ -1,0 +1,12 @@
+"""
+MCP Session Management
+
+This module handles session management and vector environment operations.
+"""
+
+from .manager import GeneralMCPVectorEnv, SessionManager
+
+__all__ = [
+    "SessionManager",
+    "GeneralMCPVectorEnv",
+]
